Using Syntax API and refactored GenericParameterDataFlow test into 2
