+++ test/Mono.Linker.Tests.Cases/DataFlow/GenericParameterDataFlow.cs
@@ -13,682 +13,9 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 	{
 		public static void Main ()
 		{
-			TestSingleGenericParameterOnType ();
-			TestMultipleGenericParametersOnType ();
-			TestBaseTypeGenericRequirements ();
-			TestDeepNestedTypesWithGenerics ();
-			TestInterfaceTypeGenericRequirements ();
-			TestTypeGenericRequirementsOnMembers ();
-			TestPartialInstantiationTypes ();
-
-			TestSingleGenericParameterOnMethod ();
-			TestMultipleGenericParametersOnMethod ();
-			TestMethodGenericParametersViaInheritance ();
-
 			MakeGenericType.Test ();
 			MakeGenericMethod.Test ();
-
-			TestNewConstraintSatisfiesParameterlessConstructor<object> ();
-			TestStructConstraintSatisfiesParameterlessConstructor<TestStruct> ();
-			TestUnmanagedConstraintSatisfiesParameterlessConstructor<byte> ();
-
-			TestGenericParameterFlowsToField ();
-			TestGenericParameterFlowsToReturnValue ();
 		}
-
-		static void TestSingleGenericParameterOnType ()
-		{
-			TypeRequiresNothing<TestType>.Test ();
-			TypeRequiresPublicFields<TestType>.Test ();
-			TypeRequiresPublicMethods<TestType>.Test ();
-			TypeRequiresPublicFieldsPassThrough<TestType>.Test ();
-			TypeRequiresNothingPassThrough<TestType>.Test ();
-		}
-
-		static void TestGenericParameterFlowsToField ()
-		{
-			TypeRequiresPublicFields<TestType>.TestFields ();
-		}
-
-		static void TestGenericParameterFlowsToReturnValue ()
-		{
-			_ = TypeRequiresPublicFields<TestType>.ReturnRequiresPublicFields ();
-			_ = TypeRequiresPublicFields<TestType>.ReturnRequiresPublicMethods ();
-			_ = TypeRequiresPublicFields<TestType>.ReturnRequiresNothing ();
-		}
-
-		[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)]
-		static Type FieldRequiresPublicFields;
-
-		[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)]
-		static Type FieldRequiresPublicMethods;
-
-		static Type FieldRequiresNothing;
-
-		class TypeRequiresPublicFields<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T>
-		{
-			[ExpectedWarning ("IL2087", "'" + nameof (T) + "'", nameof (TypeRequiresPublicFields<T>), nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-			public static void Test ()
-			{
-				typeof (T).RequiresPublicFields ();
-				typeof (T).RequiresPublicMethods ();
-				typeof (T).RequiresNone ();
-			}
-
-			[ExpectedWarning ("IL2089", "'" + nameof (T) + "'", nameof (TypeRequiresPublicFields<T>), nameof (FieldRequiresPublicMethods))]
-			public static void TestFields ()
-			{
-				FieldRequiresPublicFields = typeof (T);
-				FieldRequiresPublicMethods = typeof (T);
-				FieldRequiresNothing = typeof (T);
-			}
-
-
-			[return: DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)]
-			public static Type ReturnRequiresPublicFields ()
-			{
-				return typeof (T);
-			}
-
-
-			[ExpectedWarning ("IL2088", "'" + nameof (T) + "'", nameof (TypeRequiresPublicFields<T>), nameof (ReturnRequiresPublicMethods))]
-			[return: DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)]
-			public static Type ReturnRequiresPublicMethods ()
-			{
-				return typeof (T);
-			}
-
-			public static Type ReturnRequiresNothing ()
-			{
-				return typeof (T);
-			}
-		}
-
-		class TypeRequiresPublicMethods<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] T>
-		{
-			[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-			public static void Test ()
-			{
-				typeof (T).RequiresPublicFields ();
-				typeof (T).RequiresPublicMethods ();
-				typeof (T).RequiresNone ();
-			}
-		}
-
-		class TypeRequiresNothing<T>
-		{
-			[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-			[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-			public static void Test ()
-			{
-				typeof (T).RequiresPublicFields ();
-				typeof (T).RequiresPublicMethods ();
-				typeof (T).RequiresNone ();
-			}
-		}
-
-		class TypeRequiresPublicFieldsPassThrough<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TSource>
-		{
-			[ExpectedWarning ("IL2091", nameof (TSource),
-					"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeRequiresPublicFieldsPassThrough<TSource>",
-					"T",
-					"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeRequiresPublicMethods<T>")]
-			public static void Test ()
-			{
-				TypeRequiresPublicFields<TSource>.Test ();
-				TypeRequiresPublicMethods<TSource>.Test ();
-				TypeRequiresNothing<TSource>.Test ();
-			}
-		}
-
-		class TypeRequiresNothingPassThrough<T>
-		{
-			[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicFields<T>))]
-			[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicMethods<T>))]
-			public static void Test ()
-			{
-				TypeRequiresPublicFields<T>.Test ();
-				TypeRequiresPublicMethods<T>.Test ();
-				TypeRequiresNothing<T>.Test ();
-			}
-		}
-
-		static void TestMultipleGenericParametersOnType ()
-		{
-			MultipleTypesWithDifferentRequirements<TestType, TestType, TestType, TestType>.TestMultiple ();
-			MultipleTypesWithDifferentRequirements<TestType, TestType, TestType, TestType>.TestFields ();
-			MultipleTypesWithDifferentRequirements<TestType, TestType, TestType, TestType>.TestMethods ();
-			MultipleTypesWithDifferentRequirements<TestType, TestType, TestType, TestType>.TestBoth ();
-			MultipleTypesWithDifferentRequirements<TestType, TestType, TestType, TestType>.TestNothing ();
-		}
-
-		class MultipleTypesWithDifferentRequirements<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields | DynamicallyAccessedMemberTypes.PublicMethods)] TBoth,
-			TNothing>
-		{
-			public static void TestMultiple ()
-			{
-				typeof (TFields).RequiresPublicFields ();
-				typeof (TMethods).RequiresPublicMethods ();
-				typeof (TBoth).RequiresPublicFields ();
-				typeof (TBoth).RequiresPublicMethods ();
-				typeof (TFields).RequiresNone ();
-				typeof (TMethods).RequiresNone ();
-				typeof (TBoth).RequiresNone ();
-				typeof (TNothing).RequiresNone ();
-			}
-
-			[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-			public static void TestFields ()
-			{
-				typeof (TFields).RequiresPublicFields ();
-				typeof (TFields).RequiresPublicMethods ();
-				typeof (TFields).RequiresNone ();
-			}
-
-			[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-			public static void TestMethods ()
-			{
-				typeof (TMethods).RequiresPublicFields ();
-				typeof (TMethods).RequiresPublicMethods ();
-				typeof (TMethods).RequiresNone ();
-			}
-
-			public static void TestBoth ()
-			{
-				typeof (TBoth).RequiresPublicFields ();
-				typeof (TBoth).RequiresPublicMethods ();
-				typeof (TBoth).RequiresNone ();
-			}
-
-			[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-			[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-			public static void TestNothing ()
-			{
-				typeof (TNothing).RequiresPublicFields ();
-				typeof (TNothing).RequiresPublicMethods ();
-				typeof (TNothing).RequiresNone ();
-			}
-		}
-
-		static void TestBaseTypeGenericRequirements ()
-		{
-			new DerivedTypeWithInstantiatedGenericOnBase ();
-			new DerivedTypeWithInstantiationOverSelfOnBase ();
-			new DerivedTypeWithOpenGenericOnBase<TestType> ();
-			new DerivedTypeWithOpenGenericOnBaseWithRequirements<TestType> ();
-		}
-
-		class GenericBaseTypeWithRequirements<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T>
-		{
-			public GenericBaseTypeWithRequirements ()
-			{
-				typeof (T).RequiresPublicFields ();
-			}
-		}
-
-		class DerivedTypeWithInstantiatedGenericOnBase : GenericBaseTypeWithRequirements<TestType>
-		{
-		}
-
-		class GenericBaseTypeWithRequiresAll<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.All)] T>
-		{
-		}
-
-		class DerivedTypeWithInstantiationOverSelfOnBase : GenericBaseTypeWithRequirements<DerivedTypeWithInstantiationOverSelfOnBase>
-		{
-		}
-
-		[ExpectedWarning ("IL2091", nameof (GenericBaseTypeWithRequirements<T>))]
-		class DerivedTypeWithOpenGenericOnBase<T> : GenericBaseTypeWithRequirements<T>
-		{
-			[ExpectedWarning ("IL2091", nameof (GenericBaseTypeWithRequirements<T>))]
-			public DerivedTypeWithOpenGenericOnBase () { }
-		}
-
-		class DerivedTypeWithOpenGenericOnBaseWithRequirements<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T>
-			: GenericBaseTypeWithRequirements<T>
-		{
-		}
-
-		static void TestInterfaceTypeGenericRequirements ()
-		{
-			IGenericInterfaceTypeWithRequirements<TestType> instance = new InterfaceImplementationTypeWithInstantiatedGenericOnBase ();
-			new InterfaceImplementationTypeWithInstantiationOverSelfOnBase ();
-			new InterfaceImplementationTypeWithOpenGenericOnBase<TestType> ();
-			new InterfaceImplementationTypeWithOpenGenericOnBaseWithRequirements<TestType> ();
-		}
-
-		interface IGenericInterfaceTypeWithRequirements<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T>
-		{
-		}
-
-		class InterfaceImplementationTypeWithInstantiatedGenericOnBase : IGenericInterfaceTypeWithRequirements<TestType>
-		{
-		}
-
-		interface IGenericInterfaceTypeWithRequiresAll<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.All)] T>
-		{
-		}
-
-		class InterfaceImplementationTypeWithInstantiationOverSelfOnBase : IGenericInterfaceTypeWithRequiresAll<InterfaceImplementationTypeWithInstantiationOverSelfOnBase>
-		{
-		}
-
-		[ExpectedWarning ("IL2091", nameof (IGenericInterfaceTypeWithRequirements<T>))]
-		class InterfaceImplementationTypeWithOpenGenericOnBase<T> : IGenericInterfaceTypeWithRequirements<T>
-		{
-		}
-
-		class InterfaceImplementationTypeWithOpenGenericOnBaseWithRequirements<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T>
-			: IGenericInterfaceTypeWithRequirements<T>
-		{
-		}
-
-		static void TestDeepNestedTypesWithGenerics ()
-		{
-			RootTypeWithRequirements<TestType>.InnerTypeWithNoAddedGenerics.TestAccess ();
-		}
-
-		class RootTypeWithRequirements<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TRoot>
-		{
-			public class InnerTypeWithNoAddedGenerics
-			{
-				// The message is not ideal since we report the TRoot to come from RootTypeWithRequirements/InnerTypeWIthNoAddedGenerics
-				// while it originates on RootTypeWithRequirements, but it's correct from IL's point of view.
-				[ExpectedWarning ("IL2087", nameof (TRoot),
-						"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.RootTypeWithRequirements<TRoot>.InnerTypeWithNoAddedGenerics",
-						"type",
-						"DataFlowTypeExtensions.RequiresPublicMethods(Type)")]
-				public static void TestAccess ()
-				{
-					typeof (TRoot).RequiresPublicFields ();
-					typeof (TRoot).RequiresPublicMethods ();
-				}
-			}
-		}
-
-		static void TestTypeGenericRequirementsOnMembers ()
-		{
-			// Basically just root everything we need to test
-			var instance = new TypeGenericRequirementsOnMembers<TestType> ();
-
-			_ = instance.PublicFieldsField;
-			_ = instance.PublicMethodsField;
-
-			_ = instance.PublicFieldsProperty;
-			instance.PublicFieldsProperty = null;
-			_ = instance.PublicMethodsProperty;
-			instance.PublicMethodsProperty = null;
-
-			instance.PublicFieldsMethodParameter (null);
-			instance.PublicMethodsMethodParameter (null);
-
-			instance.PublicFieldsMethodReturnValue ();
-			instance.PublicMethodsMethodReturnValue ();
-
-			instance.PublicFieldsMethodLocalVariable ();
-			instance.PublicMethodsMethodLocalVariable ();
-		}
-
-		class TypeGenericRequirementsOnMembers<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TOuter>
-		{
-			public TypeRequiresPublicFields<TOuter> PublicFieldsField;
-
-			[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicMethods<TOuter>))]
-			public TypeRequiresPublicMethods<TOuter> PublicMethodsField;
-
-			public TypeRequiresPublicFields<TOuter> PublicFieldsProperty {
-				get;
-				set;
-			}
-
-			public TypeRequiresPublicMethods<TOuter> PublicMethodsProperty {
-				[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicMethods<TOuter>))]
-				get => null;
-				[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicMethods<TOuter>))]
-				set { }
-			}
-
-			public void PublicFieldsMethodParameter (TypeRequiresPublicFields<TOuter> param) { }
-			[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicMethods<TOuter>))]
-			public void PublicMethodsMethodParameter (TypeRequiresPublicMethods<TOuter> param) { }
-
-			public TypeRequiresPublicFields<TOuter> PublicFieldsMethodReturnValue () { return null; }
-			[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicMethods<TOuter>))]
-			public TypeRequiresPublicMethods<TOuter> PublicMethodsMethodReturnValue () { return null; }
-
-			public void PublicFieldsMethodLocalVariable ()
-			{
-				TypeRequiresPublicFields<TOuter> t = null;
-			}
-
-			[ExpectedWarning ("IL2091", nameof (TypeRequiresPublicMethods<TOuter>))]
-			public void PublicMethodsMethodLocalVariable ()
-			{
-				TypeRequiresPublicMethods<TOuter> t = null;
-			}
-		}
-
-		static void TestPartialInstantiationTypes ()
-		{
-			_ = new PartialyInstantiatedFields<TestType> ();
-			_ = new FullyInstantiatedOverPartiallyInstantiatedFields ();
-			_ = new PartialyInstantiatedMethods<TestType> ();
-			_ = new FullyInstantiatedOverPartiallyInstantiatedMethods ();
-		}
-
-		class BaseForPartialInstantiation<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods>
-		{
-		}
-
-		class PartialyInstantiatedFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TOuter>
-			: BaseForPartialInstantiation<TOuter, TestType>
-		{
-		}
-
-		class FullyInstantiatedOverPartiallyInstantiatedFields
-			: PartialyInstantiatedFields<TestType>
-		{
-		}
-
-		[ExpectedWarning ("IL2091", nameof (BaseForPartialInstantiation<TestType, TOuter>), "'TMethods'")]
-		class PartialyInstantiatedMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TOuter>
-			: BaseForPartialInstantiation<TestType, TOuter>
-		{
-			[ExpectedWarning ("IL2091", nameof (BaseForPartialInstantiation<TestType, TOuter>), "'TMethods'")]
-			public PartialyInstantiatedMethods () { }
-		}
-
-		class FullyInstantiatedOverPartiallyInstantiatedMethods
-			: PartialyInstantiatedMethods<TestType>
-		{
-		}
-
-		static void TestSingleGenericParameterOnMethod ()
-		{
-			MethodRequiresPublicFields<TestType> ();
-			MethodRequiresPublicMethods<TestType> ();
-			MethodRequiresNothing<TestType> ();
-			MethodRequiresPublicFieldsPassThrough<TestType> ();
-			MethodRequiresNothingPassThrough<TestType> ();
-		}
-
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-		static void MethodRequiresPublicFields<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ()
-		{
-			typeof (T).RequiresPublicFields ();
-			typeof (T).RequiresPublicMethods ();
-			typeof (T).RequiresNone ();
-		}
-
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-		static void MethodRequiresPublicMethods<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] T> ()
-		{
-			typeof (T).RequiresPublicFields ();
-			typeof (T).RequiresPublicMethods ();
-			typeof (T).RequiresNone ();
-		}
-
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-		static void MethodRequiresNothing<T> ()
-		{
-			typeof (T).RequiresPublicFields ();
-			typeof (T).RequiresPublicMethods ();
-			typeof (T).RequiresNone ();
-		}
-
-		[ExpectedWarning ("IL2091", nameof (MethodRequiresPublicMethods), "'T'")]
-		static void MethodRequiresPublicFieldsPassThrough<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ()
-		{
-			MethodRequiresPublicFields<T> ();
-			MethodRequiresPublicMethods<T> ();
-			MethodRequiresNothing<T> ();
-		}
-
-		[ExpectedWarning ("IL2091", nameof (MethodRequiresPublicFields), "'T'")]
-		[ExpectedWarning ("IL2091", nameof (MethodRequiresPublicMethods), "'T'")]
-		static void MethodRequiresNothingPassThrough<T> ()
-		{
-			MethodRequiresPublicFields<T> ();
-			MethodRequiresPublicMethods<T> ();
-			MethodRequiresNothing<T> ();
-		}
-
-		static void TestMethodGenericParametersViaInheritance ()
-		{
-			TypeWithInstantiatedGenericMethodViaGenericParameter<TestType>.StaticRequiresPublicFields<TestType> ();
-			TypeWithInstantiatedGenericMethodViaGenericParameter<TestType>.StaticRequiresPublicFieldsNonGeneric ();
-
-			TypeWithInstantiatedGenericMethodViaGenericParameter<TestType>.StaticPartialInstantiation ();
-			TypeWithInstantiatedGenericMethodViaGenericParameter<TestType>.StaticPartialInstantiationUnrecognized ();
-
-			var instance = new TypeWithInstantiatedGenericMethodViaGenericParameter<TestType> ();
-
-			instance.InstanceRequiresPublicFields<TestType> ();
-			instance.InstanceRequiresPublicFieldsNonGeneric ();
-
-			instance.VirtualRequiresPublicFields<TestType> ();
-			instance.VirtualRequiresPublicMethods<TestType> ();
-
-			instance.CallInterface ();
-
-			IInterfaceWithGenericMethod interfaceInstance = (IInterfaceWithGenericMethod) instance;
-			interfaceInstance.InterfaceRequiresPublicFields<TestType> ();
-			interfaceInstance.InterfaceRequiresPublicMethods<TestType> ();
-		}
-
-		class BaseTypeWithGenericMethod
-		{
-			public static void StaticRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ()
-				=> typeof (T).RequiresPublicFields ();
-			public void InstanceRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ()
-				=> typeof (T).RequiresPublicFields ();
-			public virtual void VirtualRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ()
-				=> typeof (T).RequiresPublicFields ();
-
-			public static void StaticRequiresPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] T> ()
-				=> typeof (T).RequiresPublicMethods ();
-			public void InstanceRequiresPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)]T> ()
-				=> typeof (T).RequiresPublicMethods ();
-			public virtual void VirtualRequiresPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)]T> ()
-				=> typeof (T).RequiresPublicMethods ();
-
-			public static void StaticRequiresMultipleGenericParams<
-				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-				[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods> ()
-			{
-				typeof (TFields).RequiresPublicFields ();
-				typeof (TMethods).RequiresPublicMethods ();
-			}
-		}
-
-		interface IInterfaceWithGenericMethod
-		{
-			void InterfaceRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ();
-			void InterfaceRequiresPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] T> ();
-		}
-
-		class TypeWithInstantiatedGenericMethodViaGenericParameter<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TOuter>
-			: BaseTypeWithGenericMethod, IInterfaceWithGenericMethod
-		{
-			[ExpectedWarning ("IL2091",
-				"'TInner'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeWithInstantiatedGenericMethodViaGenericParameter<TOuter>.StaticRequiresPublicFields<TInner>()",
-				"'T'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.BaseTypeWithGenericMethod.StaticRequiresPublicMethods<T>()")]
-			public static void StaticRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TInner> ()
-			{
-				StaticRequiresPublicFields<TInner> ();
-				StaticRequiresPublicMethods<TInner> ();
-			}
-
-			[ExpectedWarning ("IL2091",
-				"'TOuter'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeWithInstantiatedGenericMethodViaGenericParameter<TOuter>",
-				"'T'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.BaseTypeWithGenericMethod.StaticRequiresPublicMethods<T>()")]
-			public static void StaticRequiresPublicFieldsNonGeneric ()
-			{
-				StaticRequiresPublicFields<TOuter> ();
-				StaticRequiresPublicMethods<TOuter> ();
-			}
-
-			public static void StaticPartialInstantiation ()
-			{
-				StaticRequiresMultipleGenericParams<TOuter, TestType> ();
-			}
-
-			[ExpectedWarning ("IL2091",
-				"'TOuter'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeWithInstantiatedGenericMethodViaGenericParameter<TOuter>",
-				"'TMethods'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.BaseTypeWithGenericMethod.StaticRequiresMultipleGenericParams<TFields,TMethods>()")]
-			public static void StaticPartialInstantiationUnrecognized ()
-			{
-				StaticRequiresMultipleGenericParams<TestType, TOuter> ();
-			}
-
-			[ExpectedWarning ("IL2091",
-				"'TInner'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeWithInstantiatedGenericMethodViaGenericParameter<TOuter>.InstanceRequiresPublicFields<TInner>()",
-				"'T'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.BaseTypeWithGenericMethod.InstanceRequiresPublicMethods<T>()")]
-			public void InstanceRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TInner> ()
-			{
-				InstanceRequiresPublicFields<TInner> ();
-				InstanceRequiresPublicMethods<TInner> ();
-			}
-
-			[ExpectedWarning ("IL2091",
-				"'TOuter'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeWithInstantiatedGenericMethodViaGenericParameter<TOuter>",
-				"'T'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.BaseTypeWithGenericMethod.InstanceRequiresPublicMethods<T>()")]
-			public void InstanceRequiresPublicFieldsNonGeneric ()
-			{
-				InstanceRequiresPublicFields<TOuter> ();
-				InstanceRequiresPublicMethods<TOuter> ();
-			}
-
-			public override void VirtualRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ()
-			{
-				typeof (T).RequiresPublicFields ();
-			}
-
-			public override void VirtualRequiresPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] T> ()
-			{
-				typeof (T).RequiresPublicMethods ();
-			}
-
-			public void InterfaceRequiresPublicFields<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] T> ()
-			{
-				typeof (T).RequiresPublicFields (); ;
-			}
-
-			public void InterfaceRequiresPublicMethods<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] T> ()
-			{
-				typeof (T).RequiresPublicMethods ();
-			}
-
-			[ExpectedWarning ("IL2091",
-				"'TOuter'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.TypeWithInstantiatedGenericMethodViaGenericParameter<TOuter>",
-				"'T'",
-				"Mono.Linker.Tests.Cases.DataFlow.GenericParameterDataFlow.IInterfaceWithGenericMethod.InterfaceRequiresPublicMethods<T>()")]
-			public void CallInterface ()
-			{
-				IInterfaceWithGenericMethod interfaceInstance = (IInterfaceWithGenericMethod) this;
-				interfaceInstance.InterfaceRequiresPublicFields<TOuter> ();
-				interfaceInstance.InterfaceRequiresPublicMethods<TOuter> ();
-			}
-		}
-
-
-		static void TestMultipleGenericParametersOnMethod ()
-		{
-			MethodMultipleWithDifferentRequirements_TestMultiple<TestType, TestType, TestType, TestType> ();
-			MethodMultipleWithDifferentRequirements_TestFields<TestType, TestType, TestType, TestType> ();
-			MethodMultipleWithDifferentRequirements_TestMethods<TestType, TestType, TestType, TestType> ();
-			MethodMultipleWithDifferentRequirements_TestBoth<TestType, TestType, TestType, TestType> ();
-			MethodMultipleWithDifferentRequirements_TestNothing<TestType, TestType, TestType, TestType> ();
-		}
-
-		static void MethodMultipleWithDifferentRequirements_TestMultiple<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields | DynamicallyAccessedMemberTypes.PublicMethods)] TBoth,
-			TNothing> ()
-		{
-			typeof (TFields).RequiresPublicFields (); ;
-			typeof (TMethods).RequiresPublicMethods ();
-			typeof (TBoth).RequiresPublicFields (); ;
-			typeof (TBoth).RequiresPublicMethods ();
-			typeof (TFields).RequiresNone ();
-			typeof (TMethods).RequiresNone ();
-			typeof (TBoth).RequiresNone ();
-			typeof (TNothing).RequiresNone ();
-		}
-
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-		static void MethodMultipleWithDifferentRequirements_TestFields<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields | DynamicallyAccessedMemberTypes.PublicMethods)] TBoth,
-			TNothing> ()
-		{
-			typeof (TFields).RequiresPublicFields (); ;
-			typeof (TFields).RequiresPublicMethods ();
-			typeof (TFields).RequiresNone ();
-		}
-
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-		static void MethodMultipleWithDifferentRequirements_TestMethods<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields | DynamicallyAccessedMemberTypes.PublicMethods)] TBoth,
-			TNothing> ()
-		{
-			typeof (TMethods).RequiresPublicFields ();
-			typeof (TMethods).RequiresPublicMethods ();
-			typeof (TMethods).RequiresNone ();
-		}
-
-		static void MethodMultipleWithDifferentRequirements_TestBoth<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields | DynamicallyAccessedMemberTypes.PublicMethods)] TBoth,
-			TNothing> ()
-		{
-			typeof (TBoth).RequiresPublicFields ();
-			typeof (TBoth).RequiresPublicMethods ();
-			typeof (TBoth).RequiresNone ();
-		}
-
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicFields))]
-		[ExpectedWarning ("IL2087", nameof (DataFlowTypeExtensions.RequiresPublicMethods))]
-		static void MethodMultipleWithDifferentRequirements_TestNothing<
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields)] TFields,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicMethods)] TMethods,
-			[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicFields | DynamicallyAccessedMemberTypes.PublicMethods)] TBoth,
-			TNothing> ()
-		{
-			typeof (TNothing).RequiresPublicFields ();
-			typeof (TNothing).RequiresPublicMethods ();
-			typeof (TNothing).RequiresNone ();
-		}
-
 
 		class MakeGenericType
 		{
@@ -1141,25 +468,6 @@ namespace Mono.Linker.Tests.Cases.DataFlow
 			{
 				var t = new T ();
 			}
-		}
-
-		static void TestNewConstraintSatisfiesParameterlessConstructor<T> () where T : new()
-		{
-			RequiresParameterlessConstructor<T> ();
-		}
-
-		static void TestStructConstraintSatisfiesParameterlessConstructor<T> () where T : struct
-		{
-			RequiresParameterlessConstructor<T> ();
-		}
-
-		static void TestUnmanagedConstraintSatisfiesParameterlessConstructor<T> () where T : unmanaged
-		{
-			RequiresParameterlessConstructor<T> ();
-		}
-
-		static void RequiresParameterlessConstructor<[DynamicallyAccessedMembers (DynamicallyAccessedMemberTypes.PublicParameterlessConstructor)] T> ()
-		{
 		}
 
 		public class TestType
